Updated Infection and refactored mods to use new LGDUtilities package
Finished work on the INFECTION gametype, including options menus, weapons (a headshot capable enforcer, and a flak cannon with ONLY a shotgun primary fire) and custom sounds.

Fixed a handful of bugs with HUD rendering for the IndicatorHUD and HUDHelper, as well as the radar and Headhunter HUD.

Refactored utilities / helpers and various others to be in LGDUtilities package, instead of being located in HeadHunter.

Updated new recorded sounds for the HaloAnnouncer package.

Upadated TestBox-Large mapfile to use new actors in LGDUtilities.

Fixed issues with the Weapon/PawnHelper and changing weapons.

Fixed issues with RadarHUD scale and positioning.

Fixed issues with LinkedList and added some extra functions.

Fixed various issues with VectorHelper.

Various other bugs and improvements.....

InventoryToolbelt STILL needs work. This hasn't been completed.